Fix unnecessary delay to starting a sequential download.
When the downloader is sequential, there is no need to make a request to
the server to check if the server supports parallel downloads. This
results in an unnecessary delay to starting the request.

This modification adds only 6 lines. The getRequestSupportedFileDownloaderTypes
on HttpUrlConnectionDownloader and OkHttpDownloader simply checks if the
downloader is sequential. If it is, then the sequential downloader type
is returned immediately without querying the server for the request.